move extractors de raiz para camada transform
