+++ etl/transform/standardizer.py
@@ -2,7 +2,7 @@ import logging
 from typing import Any
 
 from etl.indexing.contracts import BronzeDocument, SilverDocument
-from etl.extractors import (
+from etl.transform.extractors import (
     display_name,
     first_value,
     match_key,
